Improve input Validations, search functionality and test DRY principles

+++ app/events/views.py
@@ -2,7 +2,7 @@
 
 from flask.views import MethodView
 from flask import make_response, request, jsonify
-from app.models import User, Events
+from app.models import User, Events, EventCategory, BlacklistToken
 
 from . import events_blueprint
 
@@ -16,19 +16,26 @@ class AllEventsView(MethodView):
 
         # GET all the events for this user
         events = Events.query
+        categories = EventCategory.get__all_categories()
         page = request.args.get('page', default=1, type=int)
         limit = request.args.get('limit', default=10, type=int)
 
         args = {}
-        potential_search = ['title', 'location', 'category']
+        potential_search = ['event_category']
 
         for search_res in potential_search:
             var = request.args.get(search_res)
             if var:
                 args.update({search_res:var})
 
-        if args:
-            events = Events.query.filter_by(**args)
+        arr = ['title', 'location']
+        categ = request.args.get('event_category')
+        if categ:
+            events = events.filter_by(event_category=categ)
+        for element in arr:
+            val = request.args.get(element)
+            if val:
+                events = events.filter(getattr(Events, element).ilike('%{}%'.format(val)))
 
         event_page = events.paginate(page, limit, False).items
 
@@ -42,8 +49,8 @@ class AllEventsView(MethodView):
                 'time': event.time,
                 'date': event.date,
                 'description': event.description,
-                'category':event.category,
-                'image_url':event.image_url
+                'image_url':event.image_url,
+                'event_category':event.event_category
             }
             results.append(obj)
 
@@ -64,6 +71,9 @@ class SingleEventView(MethodView):
     def get(event_id):
         """Handle getting a single event by id"""
         event = Events.query.filter_by(id=event_id).first_or_404()
+        categories = EventCategory.get__all_categories()
+        for category in categories:
+            event.event_category = category.category_name
         response = jsonify({
             'id': event.id,
             'title': event.title,
@@ -71,9 +81,9 @@ class SingleEventView(MethodView):
             'time': event.time,
             'date': event.date,
             'description': event.description,
-            'category':event.category,
             'image_url':event.image_url,
-            'created_by': event.created_by
+            'created_by': event.created_by,
+            'event_category':event.event_category
         })
         return make_response(response), 200
 
@@ -91,15 +101,19 @@ class UserEventsView(MethodView):
         if access_token:
             user_id = User.decode_token(access_token)
             if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
+                # checks if the user_id has a valid token or contains the user id
                 # Go ahead and handle the request, the user is authed
 
                 args = {}
                 event_models = [
-                    'title', 'location', 'category', 'time', 'date', 'description', 'image_url'
+                    'title', 'location', 'time', 'date',
+                    'description', 'image_url', 'event_category'
                 ]
 
                 for event_res in event_models:
-                    var = str(request.data.get(event_res, ''))
+                    var = str(request.data.get(event_res, '').capitalize())
+                    var = var.strip(' \t\n\r')
                     if not var:
                         response = {
                             "message":'{} missing'.format(event_res)
@@ -125,12 +139,20 @@ class UserEventsView(MethodView):
                     'time': event.time,
                     'date': event.date,
                     'description': event.description,
-                    'category':event.category,
                     'image_url':event.image_url,
-                    'created_by': user_id
+                    'created_by': user_id,
+                    'event_category': event.event_category,
+
                 })
 
                 return make_response(response), 201
+
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
+
         else:
             # user is not legit, so the payload is an error message
             message = user_id
@@ -149,11 +171,20 @@ class UserEventsView(MethodView):
         if access_token:
             user_id = User.decode_token(access_token)
             if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
                 # get all the events for this user
+                categories = EventCategory.get__all_categories()
+                page = request.args.get('page', default=1, type=int)
+                limit = request.args.get('limit', default=10, type=int)
+
                 events = Events.get_all_user(user_id)
+
+                event_page = events.paginate(page, limit, False).items
                 results = []
 
-                for event in events:
+                for event in event_page:
+                    for category in categories:
+                        event.event_category = category.category_name
                     obj = {
                         'id': event.id,
                         'title': event.title,
@@ -161,12 +192,24 @@ class UserEventsView(MethodView):
                         'time': event.time,
                         'date': event.date,
                         'description': event.description,
-                        'category':event.category,
-                        'image_url':event.image_url
+                        'image_url':event.image_url,
+                        'event_category':event.event_category
                     }
                     results.append(obj)
 
+                if not results:
+                    response = {
+                        'message': "No events found"
+                    }
+                    return make_response(jsonify(response)), 404
+
                 return make_response(jsonify(results)), 200
+
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
 
         else:
             # user is not legit, so the payload is an error message
@@ -186,12 +229,21 @@ class EventsManupilationView(MethodView):
         """Handles single event data with GET by event id."""
         auth_header = request.headers.get('Authorization')
         access_token = auth_header.split(" ")[1]
+        categories = EventCategory.get__all_categories()
 
         if access_token:
             user_id = User.decode_token(access_token)
             if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
                 # retrieve an event using it's ID
                 event = Events.query.filter_by(id=event_id).first_or_404()
+                if user_id is not event.created_by:
+                    response = {
+                        'message': 'Your do not have authorization to access this event privately'
+                    }
+                    return make_response(jsonify(response)), 401
+                for category in categories:
+                    event.event_category = category.category_name
                 response = jsonify({
                     'id': event.id,
                     'title': event.title,
@@ -199,11 +251,17 @@ class EventsManupilationView(MethodView):
                     'time': event.time,
                     'date': event.date,
                     'description': event.description,
-                    'category':event.category,
                     'image_url':event.image_url,
-                    'created_by': event.created_by
+                    'created_by': event.created_by,
+                    'event_category':event.event_category
                 })
                 return make_response(response), 200
+
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
         else:
             # user is not legit, so the payload is an error message
             message = user_id
@@ -221,24 +279,30 @@ class EventsManupilationView(MethodView):
         if access_token:
             user_id = User.decode_token(access_token)
             if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
                 # retrieve an event using it's ID
                 event = Events.query.filter_by(id=event_id).first_or_404()
+                if user_id is not event.created_by:
+                    response = {
+                        'message': 'Your do not have authorization to access this event privately'
+                    }
+                    return make_response(jsonify(response)), 401
 
                 title = str(request.data.get('title', ''))
                 location = str(request.data.get('location', ''))
                 time = str(request.data.get('time', ''))
                 date = str(request.data.get('date', ''))
                 description = str(request.data.get('description', ''))
-                category = str(request.data.get('category', ''))
                 image_url = str(request.data.get('image_url', ''))
+                event_category = str(request.data.get('event_category', ''))
 
                 event.title = title
                 event.location = location
                 event.time = time
                 event.date = date
                 event.description = description
-                event.category = category
                 event.image_url = image_url
+                event.event_category = event_category
 
                 event.save()
                 response = {
@@ -248,11 +312,17 @@ class EventsManupilationView(MethodView):
                     'time': event.time,
                     'date': event.date,
                     'description': event.description,
-                    'category':event.category,
                     'image_url':event.image_url,
-                    'created_by': event.created_by
+                    'created_by': event.created_by,
+                    'event_category': event.event_category
                 }
                 return make_response(jsonify(response)), 200
+
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
 
         else:
             # user is not legit, so the payload is an error message
@@ -271,13 +341,25 @@ class EventsManupilationView(MethodView):
         if access_token:
             user_id = User.decode_token(access_token)
             if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
                 # retrieve an event using it's ID
                 event = Events.query.filter_by(id=event_id).first_or_404()
+                if user_id is not event.created_by:
+                    response = {
+                        'message': 'Your do not have authorization to access this event privately'
+                    }
+                    return make_response(jsonify(response)), 401
 
                 event.delete()
                 return {
                     "message": "event {} deleted successfully".format(event.id)
                 }, 200
+
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
 
         else:
             # user is not legit, so the payload is an error message
@@ -301,6 +383,7 @@ class EventRsvpView(MethodView):
         if access_token:
             user_id = User.decode_token(access_token)
             if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
                 event = Events.query.filter_by(id=event_id).first_or_404()
 
                 # POST User to the RSVP
@@ -317,6 +400,12 @@ class EventRsvpView(MethodView):
                 }
                 return make_response(jsonify(response)), 200
 
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
+
         else:
             # user is not legit, so the payload is an error message
             message = user_id
@@ -326,6 +415,85 @@ class EventRsvpView(MethodView):
             return make_response(jsonify(response)), 401
 
 
+class EventCategories(MethodView):
+    """This class handles category creation and viewing"""
+
+    @staticmethod
+    def post():
+        """Handle POST request for this view. Url ---> /api/cartegory"""
+
+        auth_header = request.headers.get('Authorization')
+        access_token = auth_header.split(" ")[1]
+
+        if access_token:
+            user_id = User.decode_token(access_token)
+            if not isinstance(user_id, str):
+                BlacklistToken(token=access_token)
+                # checks if the user_id has a valid token or contains the user id
+                # Go ahead and handle the request, the user is authed
+
+                args = {}
+                event_models = ['category_name']
+
+                for event_res in event_models:
+                    var = str(request.data.get(event_res, '').capitalize())
+                    if not var:
+                        response = {
+                            "message":'{} missing'.format(event_res)
+                        }
+                        return make_response(jsonify(response)), 401
+                    args.update({event_res:var})
+
+                if EventCategory.check_category(category_name=args['category_name']):
+                    response = {
+                        "message":'Category name exists. Choose another one'
+                    }
+                    return make_response(jsonify(response)), 401
+
+                category = EventCategory(**args)
+
+                category.save()
+                response = jsonify({
+                    'id': category.id,
+                    'category_name': category.category_name
+                })
+
+                return make_response(response), 201
+
+            response_object = {
+                'status': 'fail',
+                'message': user_id
+            }
+            return make_response(jsonify(response_object)), 401
+        else:
+            # user is not legit, so the payload is an error message
+            message = user_id
+            response = {
+                'message': message
+            }
+            return make_response(jsonify(response)), 401
+
+    @staticmethod
+    def get():
+        """Handle GET request for this view. Url ---> /api/category"""
+        categories = EventCategory.get__all_categories()
+
+        results = []
+
+        for category in categories:
+            obj = {
+                'id': category.id,
+                'category_name': category.category_name
+            }
+            results.append(obj)
+
+        if not results:
+            response = {
+                'message': "Add a new category"
+            }
+            return make_response(jsonify(response)), 404
+
+        return make_response(jsonify(results)), 200
 
 # Define the API resource
 ALL_EVENTS_VIEW = AllEventsView.as_view('ALL_EVENTS_VIEW')
@@ -333,6 +501,7 @@ SINGLE_EVENT_VIEW = SingleEventView.as_view('SINGLE_EVENT_VIEW')
 USER_EVENTS_VIEW = UserEventsView.as_view('USER_EVENTS_VIEW')
 EVENT_MANUPILATION_VIEW = EventsManupilationView.as_view('EVENT_MANUPILATION_VIEW')
 EVENT_RSVP_VIEW = EventRsvpView.as_view('EVENT_RSVP_VIEW')
+EVENT_CATEGORIES_VIEW = EventCategories.as_view('EVENT_CATEGORIES_VIEW')
 
 # Define the rule for view all events url --->  /api/events/all
 # Then add the rule to the blueprint
@@ -384,3 +553,15 @@ events_blueprint.add_url_rule(
     '/api/events/<int:event_id>/rsvp',
     view_func=EVENT_RSVP_VIEW,
     methods=['POST'])
+
+# Define the rule for view all events url --->  /api/category
+# Then add the rule to the blueprint
+events_blueprint.add_url_rule(
+    '/api/category',
+    view_func=EVENT_CATEGORIES_VIEW,
+    methods=['POST'])
+
+events_blueprint.add_url_rule(
+    '/api/category',
+    view_func=EVENT_CATEGORIES_VIEW,
+    methods=['GET'])
